rearranged a couple things
