+++ transformer_lens/weight_conversion/conversion_utils/conversion_steps/weight_conversion_set.py
@@ -1,7 +1,4 @@
-import torch
-
-from .types import FIELD_SET, CONVERSION_ACTION, CONVERSION
-from transformer_lens.weight_conversion.conversion_utils.conversion_helpers import find_property
+from .types import FIELD_SET
 from .base_weight_conversion import BaseWeightConversion
 from transformer_lens.weight_conversion.conversion_utils.weight_conversion_utils import WeightConversionUtils
 
@@ -23,26 +20,6 @@ class WeightConversionSet(BaseWeightConversion):
             )
 
         return result
-
-    def process_conversion_action(self, input_value, conversion_details: CONVERSION_ACTION):
-        if isinstance(conversion_details, torch.Tensor):
-            return conversion_details
-        elif isinstance(conversion_details, str):
-            return find_property(conversion_details, input_value)
-        else:
-            (remote_field, conversion) = conversion_details
-            return self.process_conversion(input_value, remote_field, conversion)
-            
-    def process_conversion(self, input_value, remote_field: str, conversion: CONVERSION):
-        field = find_property(remote_field, input_value)
-        if isinstance(field, WeightConversionSet):
-            result = []
-            for layer in field:
-                result.append(conversion.convert(layer))
-            return result
-
-        else:
-            return conversion.convert(field)
 
     def __repr__(self):
         conversion_string = (
